Expand discussion section
Comment:
"- The final discussion is quite poor, the authors should answer to the following questions: which are the model limitations? what is the model strong for? and when does it fail? which are the future directions of this work? Can it be extended to other languages?"
Solution:
Rewrote the Discussion section to provide a more thorough analysis of model strengths, limitations, extensibility to other languages, and future work directions, replacing the previous summary of experimental insights.