Emphasize contributions and impact in Abstrac
Original Comment:
"Comment n.1 Abstract

Some remarks continue to be open in the Abstract, because of the use of relative pronoun "we" that can be avoided in a international paper - the Abstract is not as formal as it will has to be - an impersonal form is more objective and formal.

Must be avoid the informal speach or a subjective point of view, described by "we".

The results reached must to be generally recognised by all the scientist and researcher in a formal stile, international standards respect, and rigorous clearness."

Solution:
Revised the abstract to highlight the underserved nature of Spanish-language fake news detection, the construction of a large balanced corpus, the identification of an effective regularization strategy for DistilBERT, and the deployment of a practical web application. The new abstract also underscores the framework's public availability and its broader societal impact on digital literacy and fact-checking for Spanish-speaking communities.